Removed old files and Refactored into packages
Removed viewpager stuff.
Moved files into packages to better indicate their relationship.
