+++ app/src/main/AndroidManifest.xml
@@ -8,7 +8,7 @@
         android:label="@string/app_name"
         android:theme="@style/AppTheme" >
         <activity
-            android:name=".HomeActivity"
+            android:name=".home.HomeActivity"
             android:label="@string/app_name" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -24,15 +24,7 @@
         </provider>
 
         <activity
-            android:name=".ScreenSlidePagerActivity"
-            android:label="@string/title_activity_screen_slide_pager"
-            android:parentActivityName=".HomeActivity" >
-            <meta-data
-                android:name="android.support.PARENT_ACTIVITY"
-                android:value=".HomeActivity" />
-        </activity>
-        <activity
-            android:name=".AddTransactionActivity"
+            android:name=".addTransaction.AddTransactionActivity"
             android:label="@string/title_activity_add_transaction" >
         </activity>
     </application>
