+++ app/src/main/java/comp3710/csse/eng/auburn/edu/moneyapp/home/RecentTransactionsFragment.java
@@ -1,4 +1,4 @@
-package comp3710.csse.eng.auburn.edu.moneyapp;
+package comp3710.csse.eng.auburn.edu.moneyapp.home;
 
 import android.app.Activity;
 import android.net.Uri;
@@ -7,6 +7,8 @@ import android.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
+import comp3710.csse.eng.auburn.edu.moneyapp.R;
 
 
 /**
