+++ app/src/main/res/layout/activity_add_transaction.xml
@@ -6,7 +6,7 @@
                 android:paddingRight="@dimen/activity_horizontal_margin"
                 android:paddingTop="@dimen/activity_vertical_margin"
                 android:paddingBottom="@dimen/activity_vertical_margin"
-                tools:context="comp3710.csse.eng.auburn.edu.moneyapp.AddTransactionActivity">
+                tools:context="comp3710.csse.eng.auburn.edu.moneyapp.addTransaction.AddTransactionActivity">
 
     <TextView
         android:text="@string/hello_world"
