--- app/src/main/res/layout/activity_screen_slide_pager.xml
@@ -1,20 +1,0 @@
-<!-- activity_screen_slide.xml -->
-<android.support.v4.view.ViewPager
-xmlns:android="http://schemas.android.com/apk/res/android"
-xmlns:tools="http://schemas.android.com/tools"
-android:id="@+id/pager"
-android:layout_width="match_parent"
-android:layout_height="match_parent"
-tools:context="comp3710.csse.eng.auburn.edu.moneyapp.ScreenSlidePagerActivity">
-
-    <android.support.v4.view.PagerTabStrip
-        android:id="@+id/pager_title_strip"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_gravity="top"
-        android:background="#33b5e5"
-        android:textColor="#fff"
-        android:paddingTop="4dp"
-        android:paddingBottom="4dp" />
-
-</android.support.v4.view.ViewPager>
