+++ app/src/main/res/layout/fragment_choose_categories.xml
@@ -3,7 +3,7 @@
              android:layout_width="match_parent"
              android:layout_height="match_parent"
              android:orientation="vertical"
-             tools:context="comp3710.csse.eng.auburn.edu.moneyapp.ChooseCategoriesFragment">
+             tools:context="comp3710.csse.eng.auburn.edu.moneyapp.addTransaction.ChooseCategoriesFragment">
 
     <TextView
         android:layout_width="match_parent"
