--- app/src/main/res/layout/fragment_screen_slide_page.xml
@@ -1,69 +1,0 @@
-<ScrollView xmlns:android="http://schemas.android.com/apk/res/android"
-             xmlns:tools="http://schemas.android.com/tools"
-             android:layout_width="match_parent"
-             android:layout_height="match_parent"
-             tools:context="comp3710.csse.eng.auburn.edu.moneyapp.ScreenSlidePageFragment">
-
-
-    <LinearLayout
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:orientation="vertical">
-
-        <TextView
-            android:id="@+id/category_name"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:text="Category Name"/>
-
-        <TextView
-            android:id="@+id/transaction_type"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:text="Type"/>
-
-        <EditText
-            android:id="@+id/name"
-            android:layout_width="fill_parent"
-            android:layout_height="wrap_content"
-            android:hint="@string/name_hint"
-            android:inputType="text" />
-
-        <EditText
-            android:id="@+id/amount"
-            android:layout_width="fill_parent"
-            android:layout_height="wrap_content"
-            android:hint="@string/amount_hint"
-            android:inputType="numberDecimal" />
-
-        <EditText
-            android:id="@+id/time_text"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:hint="@string/pick_time"
-            android:clickable="false"
-            android:cursorVisible="false"
-            android:focusable="false"
-            android:focusableInTouchMode="false"/>
-
-        <EditText
-            android:id="@+id/date_text"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:hint="@string/pick_date"
-            android:clickable="false"
-            android:cursorVisible="false"
-            android:focusable="false"
-            android:focusableInTouchMode="false"/>
-
-        <TextView
-            android:id="@+id/accept_transaction"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="@string/accept_transaction"
-            android:clickable="false"
-            android:focusable="false"
-            android:focusableInTouchMode="false"/>
-    </LinearLayout>
-
-</ScrollView>
