+++ app/src/main/res/menu/menu_add_transaction.xml
@@ -1,7 +1,7 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
       xmlns:app="http://schemas.android.com/apk/res-auto"
       xmlns:tools="http://schemas.android.com/tools"
-      tools:context="comp3710.csse.eng.auburn.edu.moneyapp.AddTransactionActivity">
+      tools:context="comp3710.csse.eng.auburn.edu.moneyapp.addTransaction.AddTransactionActivity">
     <item android:id="@+id/action_settings"
           android:title="@string/action_settings"
           android:orderInCategory="100"
